feat(pci.ai.notebook): fix pr comments
